+++ extensions/src/ACRE2Core/PinkNoise.cpp
@@ -7,7 +7,7 @@ namespace Dsp {
 
     PinkNoise::PinkNoise() {
         
-        srand(time(NULL)); // initialize random generator
+        srand((uint32_t)time(NULL)); // initialize random generator
         clear();
     }
 
@@ -17,7 +17,7 @@ namespace Dsp {
     }
 
     float PinkNoise::tick() {
-        static const float RMI2 = 2.0 / float(RAND_MAX); // + 1.0; // change for range [0,1)
+        static const float RMI2 = 2.0f / float(RAND_MAX); // + 1.0; // change for range [0,1)
         static const float offset = A[0] + A[1] + A[2];
 
         // unrolled loop
@@ -30,6 +30,6 @@ namespace Dsp {
         return (A[0] * state[0] + A[1] * state[1] + A[2] * state[2])*RMI2 - offset;
     }
 
-    const float PinkNoise::A[PINK_NOISE_NUM_STAGES] = { 0.02109238, 0.07113478, 0.68873558 };
-    const float PinkNoise::P[PINK_NOISE_NUM_STAGES] = { 0.3190,  0.7756,  0.9613 };
+    const float PinkNoise::A[PINK_NOISE_NUM_STAGES] = { 0.02109238f, 0.07113478f, 0.68873558f };
+    const float PinkNoise::P[PINK_NOISE_NUM_STAGES] = { 0.3190f,  0.7756f,  0.9613f };
 }
